Ensure app_identifier is set in headers
This is a quick hack to actually get #45 to work...

Before this fix `@app_identifier` was always nil in the Savon.client
configuration block, as `self` was not our `Economic::Endpoint`
instance, meaning the headers would never get set. Derp.

I would like to clean this up before submitting a PR to
substancelabs/rconomic